+++ src/LGlobals.h
@@ -19,7 +19,7 @@ const int SCREEN_FPS = 60;
 const int SCREEN_TICKS_PER_FRAME = 1000 / SCREEN_FPS;
 
 struct Circle {
-	int x, y, r;
+	float x, y, r;
 };
 
 #endif
